requirements.txt
# Core Requirements
openai>=1.12.0
langchain-openai>=0.3.16
langgraph>=0.4.3
python-dotenv>=1.1.0

# LangChain Base
langchain-core>=0.1.0
langchain>=0.3.25

# Optional (for enhanced functionality)
tiktoken>=0.5.0  # For token counting
colorama>=0.4.0  # For colored console output (optional)